move : homework 함수 분할 및 함수 이름 변경
작업 완료
함수 이름 알기 쉽게 변경 및 함수 directory 변경으로 찾기 편해짐
GoldSetting.kt -> HomeworkView.kt
다크모드 색 삭제

작업 예정
UI -> View로 명명하고 함수 이름 변경하기

메인
LazyColumn 목록과 HorizontalPager로 변경 가능하게 하기
메뉴를 floating 버튼으로 바꾸기
새로고침 스낵바 시간조절

캐릭터 상세
이미지 상단 Padding 줄이기
viewModel에서 사용되는 공통 함수 model로 빼기

장비
빈값에 빈값인 이미지보여주기
-> list에서 find로 찾고 없다면 빈값, 있다면 채우기 이런식?
(나중에) 팔찌 상중하 옵션별로 색부여

스킬
룬, 보석 정렬

코드 전체
하드 코딩된 String값 정리
공통 함수 정리 및 viewModel 정리
