+++ app/src/main/java/com/hongmyeoun/goldcalc/navigation/NavControl.kt
@@ -2,30 +2,16 @@ package com.hongmyeoun.goldcalc.navigation
 
 import android.os.Build
 import androidx.annotation.RequiresApi
-import androidx.compose.animation.Crossfade
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.hongmyeoun.goldcalc.model.roomDB.character.CharacterRepository
-import com.hongmyeoun.goldcalc.view.profile.ProfileView
-import com.hongmyeoun.goldcalc.view.common.LoadingScreen
-import com.hongmyeoun.goldcalc.view.goldCheck.setting.GoldSetting
 import com.hongmyeoun.goldcalc.view.home.HomeView
+import com.hongmyeoun.goldcalc.view.homework.HomeworkView
+import com.hongmyeoun.goldcalc.view.profile.ProfileView
 import com.hongmyeoun.goldcalc.view.search.SearchView
 import com.hongmyeoun.goldcalc.view.setting.SettingView
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.AbyssDungeonVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.CommandBossVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.EpicRaidVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.GoldSettingVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.KazerothRaidVM
-import kotlinx.coroutines.delay
 
 @RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
 @Composable
@@ -44,30 +30,11 @@ fun NavControl(characterRepository: CharacterRepository) {
         }
         composable(Screen.Homework.route) {
             val charName = it.arguments?.getString("charName") ?: "ERROR"
-            val gSVM = remember { GoldSettingVM(characterRepository, charName) }
-            val character by gSVM.character.collectAsState()
-
-            var isLoading by remember { mutableStateOf(true) }
-
-            Crossfade(
-                targetState = isLoading,
-                label = "로딩 crossfade"
-            ) { loading ->
-                if (loading) {
-                    LoadingScreen()
-                } else {
-                    val cbVM = remember { CommandBossVM(character) }
-                    val adVM = remember { AbyssDungeonVM(character) }
-                    val kzVM = remember { KazerothRaidVM(character) }
-                    val epVM = remember { EpicRaidVM(character) }
-                    GoldSetting(navController, gSVM, cbVM, adVM, kzVM, epVM)
-                }
-            }
-
-            LaunchedEffect(Unit) {
-                delay(1000)
-                isLoading = false
-            }
+            HomeworkView(
+                navController = navController,
+                characterRepository = characterRepository,
+                charName = charName
+            )
         }
         composable(Screen.Search.route) {
             SearchView(navController)
@@ -81,5 +48,4 @@ fun NavControl(characterRepository: CharacterRepository) {
             SettingView(navController)
         }
     }
-
 }
