+++ app/src/main/java/com/hongmyeoun/goldcalc/view/homework/bottomBar/Summary.kt
@@ -1,8 +1,6 @@
-package com.hongmyeoun.goldcalc.view.goldCheck.setting
+package com.hongmyeoun.goldcalc.view.homework.bottomBar
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -12,33 +10,21 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowForward
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
@@ -47,45 +33,19 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.hongmyeoun.goldcalc.model.common.ImageReturn.goldImage
 import com.hongmyeoun.goldcalc.model.common.formatWithCommas
-import com.hongmyeoun.goldcalc.model.roomDB.character.Character
-import com.hongmyeoun.goldcalc.navigation.Screen
-import com.hongmyeoun.goldcalc.ui.theme.GreenQual
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
-import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
 import com.hongmyeoun.goldcalc.view.profile.titleTextStyle
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.AbyssDungeonVM
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.CommandBossVM
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.EpicRaidVM
-import com.hongmyeoun.goldcalc.viewModel.goldCheck.GoldSettingVM
+import com.hongmyeoun.goldcalc.viewModel.goldCheck.HomeworkVM
 import com.hongmyeoun.goldcalc.viewModel.goldCheck.KazerothRaidVM
 
 @Composable
-fun GoldSettingBottomBar(
-    viewModel: GoldSettingVM,
-    cbVM: CommandBossVM,
-    adVM: AbyssDungeonVM,
-    kzVM: KazerothRaidVM,
-    epVM: EpicRaidVM,
-    navController: NavHostController
-) {
-    if (viewModel.expanded) {
-        SimpleSummary(viewModel, cbVM, adVM, kzVM, epVM, navController)
-    }
-
-    BottomBar(
-        viewModel = viewModel,
-        navController = navController
-    ) { viewModel.onDoneClick(cbVM, adVM, kzVM, epVM) }
-}
-
-@Composable
 @OptIn(ExperimentalMaterial3Api::class)
-private fun SimpleSummary(
-    viewModel: GoldSettingVM,
+fun Summary(
+    viewModel: HomeworkVM,
     cbVM: CommandBossVM,
     adVM: AbyssDungeonVM,
     kzVM: KazerothRaidVM,
@@ -119,23 +79,23 @@ private fun SimpleSummary(
             ) {
 
                 if (cbVM.valtanCheck || cbVM.biaCheck || cbVM.koukuCheck || cbVM.abreCheck || cbVM.illiCheck || cbVM.kamenCheck) {
-                    SimpleCommandRaidInfo(cbVM)
+                    Script(cbVM)
                 }
 
                 if (adVM.kayangelCheck || adVM.ivoryCheck) {
-                    SimpleAbyssDungeonInfo(adVM)
+                    Script(adVM)
                 }
 
                 if (kzVM.echiCheck) {
-                    SimpleKazerothRaidInfo(kzVM)
+                    Script(kzVM)
                 }
 
                 if (epVM.beheCheck) {
-                    SimpleEpicRaidInfo(epVM)
+                    Script(epVM)
                 }
 
                 if ((viewModel.plusGold.toIntOrNull() ?: 0) > 0) {
-                    SimpleETCInfo(viewModel)
+                    Script(viewModel)
                 }
 
                 Box(modifier = Modifier.padding(start = 16.dp, end = 16.dp, bottom = 16.dp)) {
@@ -150,26 +110,24 @@ private fun SimpleSummary(
                     )
                 }
             }
-
         }
-
     }
 }
 
 @Composable
-private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
+private fun Script(cbVM: CommandBossVM) {
     Column(
         modifier = Modifier
             .padding(16.dp)
             .border(1.dp, Color.LightGray, RoundedCornerShape(8.dp))
     ) {
-        SimpleRaidTitle(title = "군단장 레이드", totalGold = cbVM.totalGold)
+        RaidTitle(title = "군단장 레이드", totalGold = cbVM.totalGold)
         Divider()
 
         Row(
             modifier = Modifier.padding(if (cbVM.valtanCheck || cbVM.biaCheck) 16.dp else 0.dp)
         ) {
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = cbVM.valtanCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "발탄",
@@ -180,7 +138,7 @@ private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
                     Text(text = "2관문 ${cbVM.valtan.twoPhase.level} : ${cbVM.valtan.twoPhase.totalGold.formatWithCommas()} G", color = Color.White)
                 }
             )
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = cbVM.biaCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "비아키스",
@@ -194,7 +152,7 @@ private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
         Row(
             modifier = Modifier.padding(if (cbVM.koukuCheck || cbVM.abreCheck) 16.dp else 0.dp)
         ) {
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = cbVM.koukuCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "쿠크세이튼",
@@ -213,7 +171,7 @@ private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
                     )
                 }
             )
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = cbVM.abreCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "아브렐슈드",
@@ -241,7 +199,7 @@ private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
         Row(
             modifier = Modifier.padding(if (cbVM.illiCheck || cbVM.kamenCheck) 16.dp else 0.dp)
         ) {
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = cbVM.illiCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "일리아칸",
@@ -254,7 +212,7 @@ private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
                     )
                 }
             )
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = cbVM.kamenCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "카멘",
@@ -270,19 +228,19 @@ private fun SimpleCommandRaidInfo(cbVM: CommandBossVM) {
 }
 
 @Composable
-private fun SimpleAbyssDungeonInfo(adVM: AbyssDungeonVM) {
+private fun Script(adVM: AbyssDungeonVM) {
     Column(
         modifier = Modifier
             .padding(16.dp)
             .border(1.dp, Color.LightGray, RoundedCornerShape(8.dp))
     ) {
-        SimpleRaidTitle(title = "어비스 던전", totalGold = adVM.totalGold)
+        RaidTitle(title = "어비스 던전", totalGold = adVM.totalGold)
         Divider()
 
         Row(
             modifier = Modifier.padding(if (adVM.kayangelCheck || adVM.ivoryCheck) 16.dp else 0.dp)
         ) {
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = adVM.kayangelCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "카양겔",
@@ -295,7 +253,7 @@ private fun SimpleAbyssDungeonInfo(adVM: AbyssDungeonVM) {
                     )
                 }
             )
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = adVM.ivoryCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "혼돈의 상아탑",
@@ -324,19 +282,19 @@ private fun SimpleAbyssDungeonInfo(adVM: AbyssDungeonVM) {
 }
 
 @Composable
-private fun SimpleKazerothRaidInfo(kzVM: KazerothRaidVM) {
+private fun Script(kzVM: KazerothRaidVM) {
     Column(
         modifier = Modifier
             .padding(16.dp)
             .border(1.dp, Color.LightGray, RoundedCornerShape(8.dp))
     ) {
-        SimpleRaidTitle(title = "카제로스 레이드", totalGold = kzVM.totalGold)
+        RaidTitle(title = "카제로스 레이드", totalGold = kzVM.totalGold)
         Divider()
 
         Row(
             modifier = Modifier.padding(if (kzVM.echiCheck) 16.dp else 0.dp)
         ) {
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = kzVM.echiCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "에키드나",
@@ -350,19 +308,19 @@ private fun SimpleKazerothRaidInfo(kzVM: KazerothRaidVM) {
 }
 
 @Composable
-private fun SimpleEpicRaidInfo(epVM: EpicRaidVM) {
+private fun Script(epVM: EpicRaidVM) {
     Column(
         modifier = Modifier
             .padding(16.dp)
             .border(1.dp, Color.LightGray, RoundedCornerShape(8.dp))
     ) {
-        SimpleRaidTitle(title = "에픽 레이드", totalGold = epVM.totalGold)
+        RaidTitle(title = "에픽 레이드", totalGold = epVM.totalGold)
         Divider()
 
         Row(
             modifier = Modifier.padding(if (epVM.beheCheck) 16.dp else 0.dp)
         ) {
-            SimplephaseInfo(
+            PhaseInfo(
                 isCheck = epVM.beheCheck,
                 modifier = Modifier.weight(1f),
                 raidName = "베히모스",
@@ -376,13 +334,13 @@ private fun SimpleEpicRaidInfo(epVM: EpicRaidVM) {
 }
 
 @Composable
-private fun SimpleETCInfo(viewModel: GoldSettingVM) {
+private fun Script(viewModel: HomeworkVM) {
     Column(
         modifier = Modifier
             .padding(16.dp)
             .border(1.dp, Color.LightGray, RoundedCornerShape(8.dp))
     ) {
-        SimpleRaidTitle(title = "기타", totalGold = viewModel.etcGold)
+        RaidTitle(title = "기타", totalGold = viewModel.etcGold)
         Divider()
 
         Row(
@@ -434,159 +392,7 @@ private fun SimpleETCInfo(viewModel: GoldSettingVM) {
 }
 
 @Composable
-private fun BottomBar(
-    viewModel: GoldSettingVM,
-    navController: NavHostController,
-    onDoneClicked: () -> Unit
-) {
-    val character by viewModel.character.collectAsState()
-
-    val borderPadding = if (viewModel.expanded) Modifier
-        .padding(start = 16.dp, end = 8.dp, bottom = 16.dp, top = 16.dp) else Modifier.padding(start = 16.dp, end = 8.dp, bottom = 16.dp)
-
-    var dragOffsetY by remember { mutableStateOf(0f) }
-    val threshold = 100f
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(
-                color = LightGrayBG,
-                shape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)
-            )
-            .clip(shape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp))
-            .pointerInput(Unit) {
-                detectDragGestures(
-                    onDrag = { change, dragAmount ->
-                        dragOffsetY += dragAmount.y
-
-                        // 드래그된 거리가 임계값을 초과하면 expand 호출
-                        if (dragOffsetY < -threshold) {
-                            viewModel.expand()
-                        }
-                    }
-                )
-            }
-            .then(borderPadding)
-    ) {
-        DragDerivationIcon()
-
-        BottomBarTexts(character, viewModel, navController, onDoneClicked)
-    }
-}
-
-@Composable
-private fun BottomBarTexts(
-    character: Character?,
-    viewModel: GoldSettingVM,
-    navController: NavHostController,
-    onDoneClicked: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Row(
-            modifier = Modifier.weight(3f),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            BottomGoldText(beforeOrAfter = "전", gold = character?.weeklyGold ?: 0)
-            Spacer(modifier = Modifier.width(8.dp))
-
-            Icon(
-                imageVector = Icons.Default.ArrowForward,
-                tint = Color.White,
-                contentDescription = "화살표"
-            )
-            Spacer(modifier = Modifier.width(12.dp))
-
-            BottomGoldText(beforeOrAfter = "후", gold = viewModel.totalGold)
-        }
-        Row(horizontalArrangement = Arrangement.End) {
-            OutlinedButton(
-                colors = ButtonDefaults.outlinedButtonColors(
-                    contentColor = Color.White
-                ),
-                onClick = {
-                    navController.navigate(Screen.Home.route) {
-                        popUpTo(Screen.Homework.route) {
-                            inclusive = true
-                        }
-                    }
-                }
-            ) {
-                Text("취소")
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            Button(
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = GreenQual,
-                    contentColor = Color.White
-                ),
-                onClick = {
-                    onDoneClicked()
-                    navController.navigate(Screen.Home.route) {
-                        popUpTo(Screen.Homework.route) {
-                            inclusive = true
-                        }
-                    }
-                }
-            ) {
-                Text(text = "완료")
-            }
-        }
-    }
-}
-
-@Composable
-private fun DragDerivationIcon() {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 12.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .width(64.dp)
-                .height(4.dp)
-                .background(
-                    color = Color.LightGray,
-                    shape = RoundedCornerShape(16.dp)
-                )
-                .align(Alignment.Center)
-        )
-    }
-    Spacer(modifier = Modifier.height(16.dp))
-}
-
-@OptIn(ExperimentalGlideComposeApi::class)
-@Composable
-fun BottomGoldText(beforeOrAfter: String, gold: Int) {
-    Column {
-        Text(
-            text = "변경 $beforeOrAfter",
-            style = titleTextStyle(fontSize = 18.sp)
-        )
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            GlideImage(
-                modifier = Modifier.size(18.dp),
-                model = goldImage(gold),
-                contentDescription = "골드 이미지",
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = gold.formatWithCommas(),
-                style = normalTextStyle(fontSize = 14.sp)
-            )
-        }
-    }
-}
-
-@Composable
-fun SimpleRaidTitle(title: String, totalGold: Int) {
+private fun RaidTitle(title: String, totalGold: Int) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
@@ -607,7 +413,7 @@ fun SimpleRaidTitle(title: String, totalGold: Int) {
 }
 
 @Composable
-fun SimplephaseInfo(
+private fun PhaseInfo(
     isCheck: Boolean,
     modifier: Modifier,
     raidName: String,
